[INLONG-8222][Manager] change package name
