+++ inlong-manager/manager-pojo/src/main/java/org/apache/inlong/manager/pojo/sort/node/base/ExtractNodeProvider.java
@@ -18,7 +18,7 @@
 package org.apache.inlong.manager.pojo.sort.node.base;
 
 import org.apache.inlong.common.enums.DataTypeEnum;
-import org.apache.inlong.manager.common.datatype.strategy.DataTypeMappingStrategy;
+import org.apache.inlong.manager.common.fieldtype.strategy.FieldTypeMappingStrategy;
 import org.apache.inlong.manager.pojo.sort.util.FieldInfoUtils;
 import org.apache.inlong.manager.pojo.stream.StreamField;
 import org.apache.inlong.manager.pojo.stream.StreamNode;
@@ -69,15 +69,15 @@ public interface ExtractNodeProvider extends NodeProvider {
      *
      * @param streamFields The stream fields
      * @param nodeId The node id
-     * @param dataTypeMappingStrategy The data type mapping operation strategy
+     * @param fieldTypeMappingStrategy The field type mapping operation strategy
      * @return FieldInfo list
      */
     default List<FieldInfo> parseStreamFieldInfos(List<StreamField> streamFields, String nodeId,
-            DataTypeMappingStrategy dataTypeMappingStrategy) {
+            FieldTypeMappingStrategy fieldTypeMappingStrategy) {
         // Filter constant fields
         return streamFields.stream().filter(s -> Objects.isNull(s.getFieldValue()))
                 .map(streamFieldInfo -> FieldInfoUtils
-                        .parseStreamFieldInfo(streamFieldInfo, nodeId, dataTypeMappingStrategy))
+                        .parseStreamFieldInfo(streamFieldInfo, nodeId, fieldTypeMappingStrategy))
                 .collect(Collectors.toList());
     }
